+++ includes/interface-aihrus-widget.php
@@ -27,7 +27,6 @@ interface Aihrus_Widget_Interface {
 	public static function form_parts( $instance, $number );
 	public static function get_content( $instance, $widget_number );
 	public static function get_defaults();
-	public static function get_suggest( $id, $suggest_id );
 	public static function validate_settings( $instance );
 }